+++ src/stats/model/weapons.rs
@@ -29,8 +29,8 @@ pub struct WeaponsInfo {
     pub headshots: u32,
     pub headshot_percentage: f32,
     pub times_chosen: u16,
-    pub bullets_fired: usize,
-    pub bullets_hit: usize,
+    pub bullets_fired: u64,
+    pub bullets_hit: u64,
     pub created: DateTime<Utc>,
     pub last_updated: DateTime<Utc>,
 }
